patch 8.2.0315: build failure on HP-UX system
Problem:    Build failure on HP-UX system.
Solution:   Use LONG_LONG_MIN instead of LLONG_MIN.  Add type casts for switch
            statement. (John Marriott)

+++ src/json.c
@@ -194,7 +194,7 @@ json_encode_item(garray_T *gap, typval_T *val, int copyID, int options)
     switch (val->v_type)
     {
 	case VAR_BOOL:
-	    switch (val->vval.v_number)
+	    switch ((long)val->vval.v_number)
 	    {
 		case VVAL_FALSE: ga_concat(gap, (char_u *)"false"); break;
 		case VVAL_TRUE: ga_concat(gap, (char_u *)"true"); break;
@@ -202,7 +202,7 @@ json_encode_item(garray_T *gap, typval_T *val, int copyID, int options)
 	    break;
 
 	case VAR_SPECIAL:
-	    switch (val->vval.v_number)
+	    switch ((long)val->vval.v_number)
 	    {
 		case VVAL_NONE: if ((options & JSON_JS) != 0
 					     && (options & JSON_NO_NONE) == 0)
